+++ Scripts/Editor/Windows/PrefabToNestedPrefab/PrefabReplacementEditor.cs
@@ -46,7 +46,7 @@ public class PrefabReplacementEditor : EditorWindow
         }
     }
 
-    [MenuItem("Nanotale/Prefab Replacement Tool", false, 1)]
+    [MenuItem( "FishingCactus/Prefab Replacement Tool", false, 1 )]
     public static void OpenPrefabReplacement()
     {
         PrefabReplacementEditor window =
@@ -74,28 +74,31 @@ public class PrefabReplacementEditor : EditorWindow
         {
             foreach ( var go in ActiveGameObjects )
             {
-                SwapPrefabs( go, new_game_object, replace );
+                GameObject instanciated_game_object = PrefabUtility.InstantiatePrefab( new_game_object ) as GameObject;
+                instanciated_game_object.transform.SetParent( go.transform.parent );
+                instanciated_game_object.transform.position = go.transform.position;
+                instanciated_game_object.transform.localScale = go.transform.localScale;
+                instanciated_game_object.transform.rotation = go.transform.rotation;
+                if ( replace )
+                {
+                    DestroyImmediate( go );
+                }
             }
         }
         else if ( ActiveGameObject != null )
         {
-            SwapPrefabs( ActiveGameObject, new_game_object, replace );
+            GameObject instanciated_game_object = PrefabUtility.InstantiatePrefab( new_game_object ) as GameObject;
+            instanciated_game_object.transform.SetParent( ActiveGameObject.transform.parent );
+            instanciated_game_object.transform.position = ActiveGameObject.transform.position;
+            instanciated_game_object.transform.localScale = ActiveGameObject.transform.localScale;
+            instanciated_game_object.transform.rotation = ActiveGameObject.transform.rotation;
+            if ( replace )
+            {
+                DestroyImmediate( ActiveGameObject );
+            }
         }
 
         return false;
-    }
-
-    private void SwapPrefabs( GameObject old_prefab, GameObject new_prefab, bool replace )
-    {
-        GameObject instanciated_game_object = PrefabUtility.InstantiatePrefab( new_prefab ) as GameObject;
-        instanciated_game_object.transform.SetParent( old_prefab.transform.parent );
-        instanciated_game_object.transform.position = old_prefab.transform.position;
-        instanciated_game_object.transform.localScale = old_prefab.transform.localScale;
-        instanciated_game_object.transform.rotation = old_prefab.transform.rotation;
-        if ( replace )
-        {
-            DestroyImmediate( old_prefab );
-        }
     }
 
     private void SearchPrefab( string search_input, string search_label )
